backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
autobahn==25.9.1
Automat==25.4.16
billiard==4.2.2
cachetools==6.2.1
celery==5.4.0
certifi==2025.10.5
cffi==2.0.0
channels==4.0.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.3
daphne==4.0.0
defusedxml==0.7.1
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-oauth-toolkit==3.1.0
django-templated-mail==1.1.1
django-timezone-field==7.1
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
djoser==2.2.0
drf-social-oauth2==2.1.0
ecdsa==0.19.1
google-ai-generativelanguage==0.4.0
google-api-core==2.26.0
google-api-python-client==2.128.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.62.3
gunicorn==22.0.0
httplib2==0.31.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
jwcrypto==1.5.6
kombu==5.5.4
msgpack==1.1.2
oauthlib==3.3.1
packaging==25.0
pillow==10.4.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python3-openid==3.2.0
redis==5.0.7
reportlab==4.2.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.4.0
social-auth-core==4.8.1
sqlparse==0.5.3
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.6.0
zope.interface==8.0.1
